fixes to properly match the types

+++ src/dashboard.js
@@ -63,10 +63,10 @@ export async function getDashboardContent (filters) {
   const data = await query
     .select(
       '*',
-      knex.raw(`(select count(*) from users_spots us where us.spot_id = spots.id) as users_count`),
-      knex.raw(`(select count(*) from spots_schools ss where ss.spot_id = spots.id) as schools_count`),
+      knex.raw(`(select count(*)::int from users_spots us where us.spot_id = spots.id) as users_count`),
+      knex.raw(`(select count(*)::int from spots_schools ss where ss.spot_id = spots.id) as schools_count`),
       knex.raw(
-        `(select count(*) from posts p where 1 = 1
+        `(select count(*)::int from posts p where 1 = 1
         and p.owner_type = 'spots'
         and p.owner_id = spots.id
         and p.image_filename is not null
